[#173] Split iRODS 4 and 5 in tests/docker/irods_catalog_provider
This commit creates a Dockerfile, entrypoint script, and unattended JSON
setup file for an iRODS catalog service provider in the Docker Compose
project for testing. To change the iRODS version, change the name of the
Dockerfile in the build/dockerfile property for the irods-catalog-provider
service in the Compose file. The iRODS 5 files are based on irods_demo
irods-catalog-provider service build files.

This also changes the default version to test against to iRODS 5.0.1.